[마이페이지] 프로필 화면 컴포넌트 분리 및 정리
- 마이페이지 화면을 구성하는 컴포넌트들을 분리하여 ProfileContent, ProfileHeader, LocationCard, ResourcesCard, InterestsCard 파일로 재구성했습니다.
- 각 컴포넌트는 자체적으로 UI를 렌더링하며 필요한 데이터만 파라미터로 받도록 하여 가독성과 재사용성을 높였습니다.
- 로그인 여부에 따라 ProfileContent 또는 NotLoggedInContent를 표시하도록 했습니다.
- ResourcesCard 내에서 잉크 레벨에 따른 색상을 결정하는 getInkLevelColor 함수를 유지했습니다.
