+++ app/src/main/java/com/example/smartee/ui/profile/ProfileHeader.kt
@@ -1,0 +1,57 @@
+package com.example.smartee.ui.profile
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import coil3.compose.AsyncImage
+import com.google.firebase.auth.FirebaseUser
+
+@Composable
+fun ProfileHeader(
+    modifier: Modifier = Modifier,
+    currentUser: FirebaseUser?
+) {
+    Column(
+        modifier = modifier,
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        // 프로필 이미지
+        AsyncImage(
+            model = currentUser?.photoUrl ?: "https://picsum.photos/300/300",
+            contentDescription = "프로필 이미지",
+            modifier = Modifier
+                .size(120.dp)
+                .clip(CircleShape),
+            contentScale = ContentScale.Crop
+        )
+
+        Spacer(modifier = Modifier.height(16.dp))
+
+        // 사용자 이름
+        Text(
+            text = currentUser?.displayName ?: "사용자",
+            style = MaterialTheme.typography.headlineMedium,
+            fontWeight = FontWeight.Bold
+        )
+
+        Spacer(modifier = Modifier.height(8.dp))
+
+        // 이메일
+        Text(
+            text = currentUser?.email ?: "",
+            style = MaterialTheme.typography.bodyMedium,
+            color = MaterialTheme.colorScheme.onSurfaceVariant
+        )
+    }
+}
